working on #156 and #146, needs cleanup and migration
