+++ NSSDB/Resources/ResourceHypermedia.cs
@@ -10,8 +10,8 @@ namespace NSSDB.Resources
     public partial class Citation : IHypermedia {[NotMapped] public List<WIM.Resources.Link> Links { get; set; }}
     public partial class EquationError : IHypermedia {[NotMapped] public List<WIM.Resources.Link> Links { get; set; } }
     public partial class Limitation : IHypermedia {[NotMapped] public List<WIM.Resources.Link> Links { get; set; } }
-    public partial class Manager : IHypermedia {[NotMapped] public List<WIM.Resources.Link> Links { get; set; } }
-    public partial class Region : IHypermedia {[NotMapped] public List<WIM.Resources.Link> Links { get; set; } }
+    //public partial class Manager : IHypermedia {[NotMapped] public List<WIM.Resources.Link> Links { get; set; } }
+    //public partial class Region : IHypermedia {[NotMapped] public List<WIM.Resources.Link> Links { get; set; } }
     public partial class RegressionRegion : IHypermedia {[NotMapped] public List<WIM.Resources.Link> Links { get; set; } }
 
 }
